+++ packages/components/src/ConfigProvider/locales/zh-tw.ts
@@ -1,11 +1,13 @@
 import { ILocale } from '../LocaleProvider/types';
+import PickerLocale from '../../DatePicker/locales/zh-tw';
 
 const localeValues: ILocale = {
   Modal: {
-    okText: '确定',
+    okText: '確定',
     cancelText: '取消',
     justOkText: '知道了',
   },
+  DatePicker: PickerLocale,
 };
 
 export default localeValues;