+++ packages/components/src/Menu/MenuItem/MenuItem.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DefaultProps } from '../../theme';
-import { ComponentPassThrough } from '../../utils/types';
+import forwardRef from '../../utils/forwardRef';
 
 export interface MenuItemProps extends DefaultProps {
   /** Item label */
@@ -22,11 +22,8 @@ export interface MenuItemType {
   ref?: React.RefObject<HTMLButtonElement> | ((instance: HTMLButtonElement) => void);
 }
 
-export function MenuItem<T extends React.ElementType = 'button'>(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  props: ComponentPassThrough<T, MenuItemProps>
-) {
+export const MenuItem = forwardRef<MenuItemProps, 'button'>(() => {
   return null;
-}
+});
 
 MenuItem.displayName = '@kubed/component/MenuItem';