+++ packages/components/src/Progress/Progress.story.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react';
+import { Progress } from './Progress';
+
+export default {
+  title: 'Components/Progress',
+  component: Progress,
+};
+
+export const Basic = () => <Progress value={50} color="#ca2621" striped />;
+
+export const segments = () => (
+  <Progress
+    size="lg"
+    sections={[
+      { value: 30, color: 'success' },
+      { value: 40, color: 'warning' },
+      { value: 10, color: 'error' },
+    ]}
+  />
+);