+++ packages/hooks/src/index.ts
@@ -9,9 +9,10 @@ export { useLocalStorage } from './useLocalStorage';
 export { useForceUpdate } from './useForceUpdate';
 export { useClipboard } from './useClipboard';
 export { useQueue } from './useQueue';
+export { useClickOutside } from './useClickOutside';
+export { useMove } from './useMove';
+export { useHotkeys } from './useHotkeys';
 // @ts-ignore
 export { useDidUpdate } from './useDidUpdate';
 // @ts-ignore
 export { useMergedRef, mergeRefs } from './useMergedRef';
-// @ts-ignore
-export { useMove } from './useMove';