+++ packages/hooks/src/utils/randomId/randomId.ts
@@ -1,3 +1,0 @@
-export function randomId() {
-  return `kubed-${Math.random().toString(36).substr(2, 9)}`;
-}