+++ packages/components/CHANGELOG.md
@@ -1,5 +1,117 @@
 # @kubed/components
 
+## 0.0.23
+
+### Patch Changes
+
+- [`5358eca`](https://github.com/kubesphere/kube-design/commit/5358eca08e547b35870d15258e3644e4bee0ed92) Thanks [@chenz24](https://github.com/chenz24)! - 1. feat: FilterInput support simple mode.
+  2. fix: Improve nav to support more flexible style setting.
+  3. fix: Fix width setting didn't take effect in Menu.
+
+## 0.0.22
+
+### Patch Changes
+
+- [`bf37ccc`](https://github.com/kubesphere/kube-design/commit/bf37cccbf64598b67b6e599b61c9d4ed33ec8504) Thanks [@chenz24](https://github.com/chenz24)! - 1. Add hotkey hook.
+  2. Fix Banner cannot set className
+- Updated dependencies [[`bf37ccc`](https://github.com/kubesphere/kube-design/commit/bf37cccbf64598b67b6e599b61c9d4ed33ec8504)]:
+  - @kubed/hooks@0.0.4
+
+## 0.0.21
+
+### Patch Changes
+
+- [`7c7e066`](https://github.com/kubesphere/kube-design/commit/7c7e066d5a612344865f8279d39582e65c71919a) Thanks [@chenz24](https://github.com/chenz24)! - 1. fix: Omit hideOnClick if Dropdown component is control component.
+  2. feat: Add useMove hook.
+  3. Fix MenuItem cannot be render if place in react fragment
+
+## 0.0.20
+
+### Patch Changes
+
+- [`e3b1a57`](https://github.com/kubesphere/kube-design/commit/e3b1a57e6adb6480963399e30151978511d89350) Thanks [@chenz24](https://github.com/chenz24)! - 1. Add useClickOutside hook
+  2. Add FilterInput component.
+  3. Tag support append element.
+  4. Some other bugs fix.
+
+## 0.0.19
+
+### Patch Changes
+
+- [`dc41c69`](https://github.com/kubesphere/kube-design/commit/dc41c69b266e08345d7c43222bee5d1fabe10087) Thanks [@chenz24](https://github.com/chenz24)! - 1. Export Progress component.
+  2. Omit trigger if visible is set.
+
+## 0.0.18
+
+### Patch Changes
+
+- [`a7f7e6a`](https://github.com/kubesphere/kube-design/commit/a7f7e6aa089002b1c24ab650aef124a7ecc3157d) Thanks [@chenz24](https://github.com/chenz24)! - 1. Use classnames instead of clsx.
+  2. Fix package bug cause by importing dayjs locales
+
+## 0.0.17
+
+### Patch Changes
+
+- [`6fc0e96`](https://github.com/kubesphere/kube-design/commit/6fc0e96fb62e75eb9f1eb5c79347896f8c971cc1) Thanks [@chenz24](https://github.com/chenz24)! - 1. Fix import wrong 'visible' from chalk.
+  2. Improve i18n.
+
+* [`eb0b852`](https://github.com/kubesphere/kube-design/commit/eb0b8520f105ebd91d25a0b9a7b8ab3c6a86b7fc) Thanks [@chenz24](https://github.com/chenz24)! - 1. Fix import wrong 'visible' from chalk.
+  2. Improve i18n.
+
+## 0.0.16
+
+### Patch Changes
+
+- [`0f4d178`](https://github.com/kubesphere/kube-design/commit/0f4d17860c1ae0c5ebcfdcd61b3b2288df7fb877) Thanks [@chenz24](https://github.com/chenz24)! - 1. feat: Add Progress component.
+  2. fix: Set hover color for menu item.
+  3. feat: Add ActionConfirm component
+
+## 0.0.15
+
+### Patch Changes
+
+- [`c619355`](https://github.com/kubesphere/kube-design/commit/c6193554484524736c95bf263a1af1a6fdd394e8) Thanks [@chenz24](https://github.com/chenz24)! - 1. Divider support height and color.
+  2. Fix modal cannot hide header and footer.
+  3. Fix react key warning when open imperative modal.
+  4. Popover support set width and contentClassName.
+
+## 0.0.14
+
+### Patch Changes
+
+- [`81610f7`](https://github.com/kubesphere/kube-design/commit/81610f7b72e85c675dfd129bd5c676ad09c6f702) Thanks [@chenz24](https://github.com/chenz24)! - 1. Fix ModalContext default value.
+  2. Fix MenuItem error cause by 'as' props.
+  3. Fix background of doc page in dark mode.
+
+## 0.0.13
+
+### Patch Changes
+
+- [`e1fe306`](https://github.com/kubesphere/kube-design/commit/e1fe30612f48839ae6eb40522ba28dd893ccbf9f) Thanks [@chenz24](https://github.com/chenz24)! - 1. Set default size of icon to 16.
+  2. Transient props in Button for styled-component and set default radius to xl.
+- Updated dependencies [[`e1fe306`](https://github.com/kubesphere/kube-design/commit/e1fe30612f48839ae6eb40522ba28dd893ccbf9f)]:
+  - @kubed/icons@0.0.2
+
+## 0.0.12
+
+### Patch Changes
+
+- [`0cff4d4`](https://github.com/kubesphere/kube-design/commit/0cff4d4ea509b874b0a746b8a4dbaa24f7955c62) Thanks [@chenz24](https://github.com/chenz24)! - - fix: build error cause by Select export.
+
+## 0.0.11
+
+### Patch Changes
+
+- [`c8be7d3`](https://github.com/kubesphere/kube-design/commit/c8be7d31582ee31fa7c33ee89df5d36dd659ab36) Thanks [@chenz24](https://github.com/chenz24)! - - Remove Slider and AutoComplete for their wrong export.
+
+## 0.0.10
+
+### Patch Changes
+
+- [`c2610f8`](https://github.com/kubesphere/kube-design/commit/c2610f898a081e743918bc755a308ae2c2016035) Thanks [@chenz24](https://github.com/chenz24)! - 1. Add AutoComplete component.
+  2. Add Slider component.
+  3. Notify support adding title.
+
 ## 0.0.9
 
 ### Patch Changes